Add service method to explicity add a highscore to a map when playing in single player mode.

+++ src/Billapong.DataAccess/Model/GamePlay/HighScore.cs
@@ -41,6 +41,6 @@
         /// The score.
         /// </value>
         [Required]
-        public long Score { get; set; }
+        public int Score { get; set; }
     }
 }
